# scca-rest/src/main/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample.",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "${scca.rest.context-path}/property/persistent" : {
      "get" : {
        "tags" : [ "Property MGT（配置详细管理）" ],
        "summary" : "persistent properties / 获取持久化的配置信息",
        "description" : "",
        "operationId" : "propertiesFromPersistent",
        "parameters" : [ {
          "name" : "project",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "profile",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "label",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespProperties"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Property MGT（配置详细管理）" ],
        "summary" : "Save properties / 保存持久化的配置信息",
        "description" : "",
        "operationId" : "saveProperties",
        "parameters" : [ {
          "name" : "project",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "profile",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "label",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "WebRespProperties" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "object",
          "additionalProperties" : {
            "type" : "string"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
    "WebRespString" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
  }
}